+++ packages/hydrogen/src/components/LocalizationProvider/index.ts
@@ -1,2 +1,3 @@
+export {LocalizationProvider} from './LocalizationProvider.server';
 export {useCountry} from '../../hooks/useCountry/useCountry';
 export {useAvailableCountries} from '../../hooks/useAvailableCountries/useAvailableCountries';