+++ packages/hydrogen/src/foundation/useQuery/index.ts
@@ -1,2 +1,1 @@
 export {useQuery} from './hooks';
-export {QueryProvider} from './QueryProvider';