+++ scripts/generate-docs.ts
@@ -1,6 +1,6 @@
 import {resolve} from 'path';
 // eslint-disable-next-line node/no-missing-import
-import {DocsGen, Options} from '../packages/generate-docs';
+import {DocsGen, Options, Column} from '../packages/generate-docs';
 
 const ROOT = resolve('.');
 const logger = console;
@@ -10,6 +10,12 @@ async function runCliGenerator(args: Partial<Options> = {}) {
     inputRootPath: ROOT,
     packageName: 'cli',
     ...args,
+  });
+
+  const CommandsTable = await cliGenerator.table({
+    title: 'Commands',
+    description: 'The Hydrogen CLI has the following commands',
+    columns: [Column.ComponentName, Column.Description],
   });
 
   await Promise.all([
@@ -26,7 +32,7 @@ async function runCliGenerator(args: Partial<Options> = {}) {
       description: 'Command reference for the `@shopify/hydrogen-cli`',
       url: '/api/hydrogen/cli/commands/index.md',
       entry: 'commands',
-      tableColumns: ['Command', 'Description'],
+      tables: [CommandsTable],
       hidden: true,
     }),
     cliGenerator.section({
@@ -34,11 +40,16 @@ async function runCliGenerator(args: Partial<Options> = {}) {
       description: 'Create command reference from the `@shopify/hydrogen-cli`',
       url: '/api/hydrogen/cli/commands/create/index.md',
       entry: 'commands/create',
-      tableColumns: ['Command', 'Description'],
       hidden: true,
+      tables: [CommandsTable],
     }),
   ]);
 }
+
+const COMPONENTS_TABLE = {
+  title: 'Reference',
+  columns: [Column.ComponentName, Column.ComponentType, Column.Description],
+};
 
 async function runHydrogenGenerator(args: Partial<Options> = {}) {
   const generator = new DocsGen({
@@ -46,6 +57,33 @@ async function runHydrogenGenerator(args: Partial<Options> = {}) {
     packageName: 'hydrogen',
     ...args,
   });
+
+  // Tables
+  const primitiveComponentsTable = await generator.table({
+    ...COMPONENTS_TABLE,
+    description: 'Hydrogen includes the following primitive components:',
+  });
+
+  const cartComponentsTable = await generator.table({
+    ...COMPONENTS_TABLE,
+    description: 'Hydrogen includes the following cart components:',
+  });
+
+  const productAndVariantTable = await generator.table({
+    ...COMPONENTS_TABLE,
+    description:
+      'Hydrogen includes the following product and variant components:',
+  });
+
+  const localizationTable = await generator.table({
+    ...COMPONENTS_TABLE,
+    description: 'Hydrogen includes the following localization components:',
+  });
+  const globalComponentsTable = await generator.table({
+    ...COMPONENTS_TABLE,
+    description: 'Hydrogen includes the following global components:',
+  });
+
   await Promise.all([
     // Components
     generator.section({
@@ -57,9 +95,11 @@ async function runHydrogenGenerator(args: Partial<Options> = {}) {
     }),
     // Primitive
     generator.section({
-      title: 'Primitive',
+      title: 'Primitive components',
+      intro:
+        'Primitive components are the building blocks for different component types, including products, variants, and cart.',
       description:
-        'Get familiar with the Hydrogen primitive components included in Hydrogen.',
+        'Get familiar with the primitive components included in Hydrogen.',
       url: '/api/hydrogen/components/primitive/index.md',
       entry: [
         'components/ExternalVideo',
@@ -73,20 +113,26 @@ async function runHydrogenGenerator(args: Partial<Options> = {}) {
         'components/UnitPrice',
         'components/Video',
       ],
+      tables: [primitiveComponentsTable],
     }),
     // Global
     generator.section({
-      title: 'Global',
+      title: 'Global components',
+      intro:
+        'Global components are React components that relate to your entire app.',
       description:
-        'Get familiar with the Hydrogen global components included in Hydrogen.',
+        'Get familiar with the global components included in Hydrogen.',
       url: '/api/hydrogen/components/global/index.md',
-      entry: 'foundation/ShopifyProvider',
+      entry: ['foundation/ShopifyProvider'],
+      tables: [globalComponentsTable],
     }),
     // Product and variant
     generator.section({
-      title: 'Product and variant',
+      title: 'Product and variant components',
+      intro:
+        'Products are the goods, digital downloads, services, and gift cards that a merchant sells. If a product has options, like size or color, then merchants can add a variant for each combination of options. Each combination of option values for a product can be a variant for that product. For example, a t-shirt might be available for purchase in blue and green. The blue t-shirt and the green t-shirt are variants.',
       description:
-        'Get familiar with the Hydrogen product and variant components included in Hydrogen.',
+        'Get familiar with the product and variant components included in Hydrogen.',
       url: '/api/hydrogen/components/product-variant/index.md',
       entry: [
         'components/ProductDescription',
@@ -102,12 +148,15 @@ async function runHydrogenGenerator(args: Partial<Options> = {}) {
         'components/SelectedVariantShopPayButton',
         'components/SelectedVariantUnitPrice',
       ],
+      tables: [productAndVariantTable],
     }),
     // Cart
     generator.section({
-      title: 'Cart',
+      title: 'Cart components',
+      intro:
+        'A cart contains the merchandise that a customer intends to purchase and the estimated cost associated with the cart. When a customer is ready to purchase their items, they can proceed to checkout.',
       description:
-        'Get familiar with the Hydrogen cart components included in Hydrogen.',
+        'Get familiar with the cart components included in Hydrogen.',
       url: '/api/hydrogen/components/cart/index.md',
       entry: [
         'components/AddToCartButton',
@@ -126,14 +175,18 @@ async function runHydrogenGenerator(args: Partial<Options> = {}) {
         'components/CartProvider',
         'components/CartShopPayButton',
       ],
+      tables: [cartComponentsTable],
     }),
     // Localization
     generator.section({
-      title: 'Localization',
+      title: 'Localization components',
+      intro:
+        'Localization can help merchants expand their business to a global audience by creating shopping experiences in local languages and currencies.',
       description:
-        'Get familiar with the Hydrogen localization components included in Hydrogen.',
+        'Get familiar with the localization components included in Hydrogen.',
       url: '/api/hydrogen/components/localization/index.md',
-      entry: 'components/LocalizationProvider',
+      entry: ['components/LocalizationProvider'],
+      tables: [localizationTable],
     }),
 
     // Hooks